requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
cachetools>=5.5.0
requests>=2.32.0
locust>=2.32.0
pyyaml>=6.0.2
graphql-core>=3.2.5
psutil>=6.1.0
python-multipart>=0.0.18
cryptography>=44.0.0
